Application of Explainable AI
Some people might think the model you are creating is a black box, even we don't understand well enough the argument behind the prediction made by our model. Explainable AI is a method to show the relation behind features and can determine the reason for every instance prediction of the data.